move extra types to appropiate packages
